+++ integration/setup/setup.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+apptestctl bootstrap --kubeconfig="$(kind get kubeconfig)" --install-operators=false