+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+### Changed
+
+- Use `apptestctl` to install CRDs in integration tests to avoid hitting GitHub rate limits.
+
 ## [2.20.0] - 2021-12-15
 
 ### Changed